new code

+++ src/utils/cloudinary.js
@@ -1,25 +1,32 @@
-import {v2 as cloudinary } from "cloudinary"
-import fs from "fs"
-cloudinary.config({
-    cloud_name:process.env.CLOUDINARY_CLOUD_NAME,
-    api_key:process.env.CLOUDINARY_API_KEY,
-    api_secret:process.env.CLOUDINARY_API_SECRET
+import dotenv from "dotenv";
+dotenv.config({ path: ".env" });
 
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
+
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-const uploadOnCloudinary=async(localFilePath)=>{
-    try {
-        if(!localFilePath) return null
+const uploadOnCloudinary = async (localFilePath) => {
 
-        const response = await cloudinary.uploader.upload(localFilePath,{resource_type:"auto"})
+  try {
+    if (!localFilePath) return null;
 
-        console.log("File uploaded on cloudinary",response.url);
-        return response;
-        
-    } catch (error) {
-        fs.unlinkSync(localFilePath)
-        return null;
-        
-    }
-}
-export {uploadOnCloudinary}
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
+
+    fs.unlinkSync(localFilePath);
+    console.log("Upload success:", response.url);
+    return response;
+  } catch (error) {
+    console.error("Cloudinary upload failed:", error.message);
+    if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
